+++ paging-samples/src/commonMain/kotlin/ua/wwind/paging/sample/presentation/screen/MainTabsScreen.kt
@@ -6,9 +6,11 @@ import androidx.compose.material3.TabRow
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableIntStateOf
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.launch
 import ua.wwind.paging.sample.data.local.InMemoryUserLocalDataSource
 import ua.wwind.paging.sample.data.local.SharedEditableUsersStore
 import ua.wwind.paging.sample.data.repository.FakeUserRemoteDataSource
@@ -41,19 +43,24 @@ fun MainTabsScreen(
                     useMediator = false,
                     scope = scope
                 )
+                val useMediator = remember { mutableStateOf(false) }
                 UserListScreen(
                     viewModel = vm,
-                    useMediator = false,
-                    onToggleUseMediator = {},
-                    onRefresh = {},
-                    cachedCountFlow = null,
-                    lastMinSavedKeyFlow = null,
+                    useMediator = useMediator.value,
+                    onToggleUseMediator = { useMediator.value = it },
+                    onRefresh = {
+                        scope.launch {
+                            vm.clearCache()
+                        }
+                    },
+                    cachedCountFlow = local.cachedCount,
+                    lastMinSavedKeyFlow = local.lastSavedMinKey,
                     modifier = modifier
                 )
             }
 
             else -> {
-                val store = SharedEditableUsersStore(initialSize = 200)
+                val store = SharedEditableUsersStore(initialSize = 500)
                 val vm = StreamingUserListViewModel(store = store, scope = scope)
                 StreamingUserListScreen(viewModel = vm, modifier = modifier)
             }