+++ paging-samples/src/commonMain/kotlin/ua/wwind/paging/sample/presentation/viewmodel/StreamingUserListViewModel.kt
@@ -2,12 +2,14 @@ package ua.wwind.paging.sample.presentation.viewmodel
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.Flow
+import ua.wwind.paging.core.ExperimentalStreamingPagerApi
 import ua.wwind.paging.core.PagingData
 import ua.wwind.paging.core.StreamingPager
 import ua.wwind.paging.core.StreamingPagerConfig
 import ua.wwind.paging.sample.data.local.SharedEditableUsersStore
 import ua.wwind.paging.sample.domain.model.User
 
+@OptIn(ExperimentalStreamingPagerApi::class)
 class StreamingUserListViewModel(
     private val store: SharedEditableUsersStore,
     scope: CoroutineScope,